requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
channels==4.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.3
django-ckeditor==6.7.0
django-cors-headers==4.6.0
django-filter==24.3
django-jazzmin==3.0.1
django-js-asset==2.1.0
django-modeltranslation==0.18.11
django-parler==2.3
django-parler-rest==2.2
django-redis==5.4.0
django-tinymce==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
kombu==5.4.2
packaging==23.2
pillow==11.0.0
prompt_toolkit==3.0.48
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-environ==0.4.54
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.2.0
setuptools==75.1.0
six==1.16.0
sorl-thumbnail==12.11.0
sqlparse==0.5.1
typing_extensions==4.7.1
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
